HACK: Initial partially working ARM support
So far supports only ARM Linux with hardfloat ABI (armv7l-linux-gnueabihf)
and works fine for simple helloworld-style applications. For the usage
instructions, see:
   http://crystal-lang.org/docs/syntax_and_semantics/cross-compilation.html

But exceptions handling does not work correctly yet! The libunwind
stuff still needs to be debugged! So we can't build the whole Crystal
compiler for ARM and pass the spec test suite at the moment. Stay tuned.